requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
openai-whisper==20231117
python-multipart==0.0.6
pydantic==2.5.2
websockets==12.0
pinecone-client==2.2.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 